Fixed typo in SessionService.groovy.

+++ src/main/groovy/com/lorenjamison/alexa/triviaroyale/service/SessionService.groovy
@@ -16,7 +16,7 @@ class SessionService {
     static Session getSessionFromAlexaSessionAttributes(Map<String, Object> sessionAttributes) {
         Session session = new Session()
         session.with {
-            id = sessionAttributes[SessionAttributes.SESSION_ID]
+            id = sessionAttributes[SessionAttributes.SESSION_ID] as long
             status = SessionStatus.ACTIVE   //we don't need to store the session status since all incoming sessions are active
             quizId = sessionAttributes[SessionAttributes.QUIZ_ID] as long
             currentQuestionIndex = sessionAttributes[SessionAttributes.QUESTION_NUMBER] as int
